Added 3D effect with sprites
changed it so enemy/enemy, player/player, and bullet/bullet collisions are disabled (so they will pass through each other)
3D effect works by having objects that are lower down appear overtop higher objects
