+++ Assets/Scenes/Scene.unity
@@ -420,7 +420,7 @@ GameObject:
   - component: {fileID: 791496637}
   - component: {fileID: 791496636}
   - component: {fileID: 791496635}
-  m_Layer: 0
+  m_Layer: 5
   m_Name: EventSystem
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -22167,7 +22167,7 @@ SpriteRenderer:
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: -1011493385
-  m_SortingLayer: 4
+  m_SortingLayer: 2
   m_SortingOrder: 0
   m_Sprite: {fileID: 21300000, guid: 7dbc73b5bd853e446a8c2535ffe2d449, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
