+++ Assets/Scripts/Controllers/ObjectController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 856f2f1e98d38034daeb9d17c38926a1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
